♻️ certif: Move creation and edition modal into session-details
