+++ certif/app/components/sessions/session-details/enrolled-candidates/candidate-details-modal-row.gjs
